Rename package arrow.mtl.test
